+++ lib/DigitalSensor/DigitalSensor.h
@@ -11,7 +11,7 @@ class DigitalSensor : public Sensor {
 private:
     uint8_t pin;
 protected:
-    uint8_t getSensorValue();
+    uint16_t getSensorValue();
 public:
     /**
      * Constructor, creates an instance of a DigitalSensor
